Move finite element code into a dedicated module for reusability
