+++ Sofa/Component/SolidMechanics/FEM/Elastic/Sofa.Component.SolidMechanics.FEM.ElasticConfig.cmake.in
@@ -6,6 +6,7 @@
 find_package(Sofa.Config QUIET REQUIRED)
 sofa_find_package(Sofa.Simulation.Core QUIET REQUIRED)
 sofa_find_package(Sofa.Component.Topology.Container.Grid QUIET REQUIRED)
+sofa_find_package(Sofa.FEM QUIET REQUIRED)
 
 if(NOT TARGET @PROJECT_NAME@)
     include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
