+++ Sofa/framework/CMakeLists.txt
@@ -12,6 +12,7 @@ sofa_add_subdirectory(library Helper Sofa.Helper ON)
 sofa_add_subdirectory(library LinearAlgebra Sofa.LinearAlgebra ON)
 sofa_add_subdirectory(library DefaultType Sofa.DefaultType ON)
 sofa_add_subdirectory(library Core Sofa.Core ON)
+sofa_add_subdirectory(library FEM Sofa.FEM ON)
 sofa_add_subdirectory(library Simulation Sofa.Simulation ON)
 sofa_add_subdirectory(library SimpleApi Sofa.SimpleApi ON)
 sofa_add_subdirectory(library Testing Sofa.Testing ON)
