+++ Sofa/framework/FEM/src/sofa/fem/FiniteElement[Triangle].h
@@ -19,19 +19,53 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#define ELASTICITY_COMPONENT_BASE_ELEMENT_LINEAR_FEM_FORCEFIELD_CPP
-#include <sofa/component/solidmechanics/fem/elastic/BaseElementLinearFEMForceField.inl>
-#include <sofa/fem/FiniteElement[all].h>
+#pragma once
+#include <sofa/fem/FiniteElement.h>
 
-namespace sofa::component::solidmechanics::fem::elastic
+namespace sofa::fem
 {
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec1Types, sofa::geometry::Edge>;
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec2Types, sofa::geometry::Edge>;
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec3Types, sofa::geometry::Edge>;
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec2Types, sofa::geometry::Triangle>;
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec3Types, sofa::geometry::Triangle>;
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec2Types, sofa::geometry::Quad>;
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec3Types, sofa::geometry::Quad>;
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec3Types, sofa::geometry::Tetrahedron>;
-template class SOFA_COMPONENT_SOLIDMECHANICS_FEM_ELASTIC_API BaseElementLinearFEMForceField<sofa::defaulttype::Vec3Types, sofa::geometry::Hexahedron>;
+
+#if !defined(SOFA_FEM_FINITE_ELEMENT_TRIANGLE_CPP)
+#include <sofa/defaulttype/VecTypes.h>
+#endif
+
+template <class DataTypes>
+struct FiniteElement<sofa::geometry::Triangle, DataTypes>
+{
+    FINITEELEMENT_HEADER(sofa::geometry::Triangle, DataTypes, 2);
+    static_assert(spatial_dimensions > 1, "Triangles cannot be defined in 1D");
+
+    constexpr static std::array<ReferenceCoord, NumberOfNodesInElement> referenceElementNodes {{
+        {0, 0},
+        {1, 0},
+        {0, 1}}};
+
+    static const sofa::type::vector<TopologyElement>& getElementSequence(sofa::core::topology::BaseMeshTopology& topology)
+    {
+        return topology.getTriangles();
+    }
+
+    static constexpr sofa::type::Mat<NumberOfNodesInElement, TopologicalDimension, Real> gradientShapeFunctions(const sofa::type::Vec<TopologicalDimension, Real>& q)
+    {
+        SOFA_UNUSED(q);
+        return {
+            {-1, -1},
+            {1, 0},
+            {0, 1}
+        };
+    }
+
+    static constexpr std::array<QuadraturePointAndWeight, 1> quadraturePoints()
+    {
+        return {
+            std::make_pair(sofa::type::Vec<TopologicalDimension, Real>(1./3., 1./3.), 1./2.)
+        };
+    }
+};
+
+#if !defined(SOFA_FEM_FINITE_ELEMENT_TRIANGLE_CPP)
+extern template struct SOFA_FEM_API FiniteElement<sofa::geometry::Triangle, sofa::defaulttype::Vec3Types>;
+extern template struct SOFA_FEM_API FiniteElement<sofa::geometry::Triangle, sofa::defaulttype::Vec2Types>;
+#endif
+
 }
